worker的管理和任务转移

+++ model/worker.go
@@ -10,8 +10,8 @@ type Worker struct {
 	/*worker机器必须要有此，不然不会加入可使用worker队列*/
 	Key    string
 
-	/*状态，服务端控制是否使用此机器, 0未加入队列, 1 加入队列中*/
-	Status int8
+	/*状态，服务端控制是否使用此机器, 0未加入队列, 1 加入队列中, 2 禁用*/
+	Status int
 
 	/*备注*/
 	Note   string
